WDS-57 add collapse transition component
