+++ packages/ui/src/components/Transitions/index.ts
@@ -1,0 +1,1 @@
+export { default as SCollapseTransition } from './SCollapseTransition.vue'
