+++ packages/ui/src/components/index.ts
@@ -15,4 +15,5 @@ export * from './ProgressBar'
 export * from './Tabs'
 export * from './Badge'
 export * from './Checkbox'
+export * from './Transitions'
 export * from './Radio'
